refactor(application_utils): extract base function
