+++ Cargo.toml
@@ -36,6 +36,7 @@ aws-sdk-s3 = { version = "=1.42.0", features = ["behavior-version-latest"] }
 axum = { version = "=0.7.5", features = ["macros", "multipart"] }
 boilermates = "=0.3.0"
 chrono = "=0.4.38"
+const-str = "=0.5.7"
 dotenvy_macro = "=0.15.7"
 enum_meta = "=0.7.0"
 nanoid = "=0.4.0"
@@ -69,5 +70,10 @@ serde = { version = "=1.0.204", features = ["derive"] }
 serde_json = "=1.0.120"
 serde_with = { version = "=3.9.0", features = ["chrono_0_4"] }
 strum = { version = "=0.26.2", features = ["derive"] }
+# FIXME: Upgrade once https://github.com/seanmonstar/reqwest/pull/1620 is merged
+reqwest = { git = "https://github.com/thomasqueirozb/reqwest", branch = "base_url", features = [
+    "json",
+    "stream",
+] }
 tracing = { version = "=0.1.40", features = ["attributes"] }
 uuid = { version = "=1.10.0", features = ["v4"], default-features = false }
