+++ apps/backend/Cargo.toml
@@ -21,7 +21,7 @@ chrono-tz = "=0.9.0"
 compile-time = "=0.2.0"
 config = { path = "../../crates/config" }
 convert_case = "=0.6.0"
-const-str = "=0.5.7"
+const-str = { workspace = true }
 csv = "=1.3.0"
 data-encoding = "=2.6.0"
 dotenvy = "=0.15.7"
@@ -46,11 +46,7 @@ paginate = "=1.1.11"
 radarr-api-rs = "=3.0.1"
 rand = "=0.9.0-alpha.2"
 regex = "=1.10.5"
-# FIXME: Upgrade once https://github.com/seanmonstar/reqwest/pull/1620 is merged
-reqwest = { git = "https://github.com/thomasqueirozb/reqwest", branch = "base_url", features = [
-    "json",
-    "stream",
-] }
+reqwest = { workspace = true }
 rust_decimal = { workspace = true }
 rust_decimal_macros = { workspace = true }
 rust_iso3166 = "=0.1.13"
