+++ apps/backend/src/providers/google_books.rs
@@ -15,8 +15,7 @@ use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use traits::{MediaProvider, MediaProviderLanguages};
 use utils::convert_date_to_year;
-
-use crate::app_utils::get_base_http_client;
+use utils::get_base_http_client;
 
 static URL: &str = "https://www.googleapis.com/books/v1/volumes/";
