+++ apps/backend/src/providers/itunes.rs
@@ -11,8 +11,7 @@ use reqwest::Client;
 use sea_orm::prelude::ChronoDateTimeUtc;
 use serde::{Deserialize, Serialize};
 use traits::{MediaProvider, MediaProviderLanguages};
-
-use crate::app_utils::get_base_http_client;
+use utils::get_base_http_client;
 
 static URL: &str = "https://itunes.apple.com/";
