+++ apps/backend/src/providers/openlibrary.rs
@@ -14,8 +14,7 @@ use scraper::{Html, Selector};
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use traits::{MediaProvider, MediaProviderLanguages};
-
-use crate::app_utils::get_base_http_client;
+use utils::get_base_http_client;
 
 static URL: &str = "https://openlibrary.org/";
 static IMAGE_BASE_URL: &str = "https://covers.openlibrary.org";
