+++ apps/backend/src/providers/vndb.rs
@@ -11,9 +11,8 @@ use reqwest::Client;
 use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use traits::{MediaProvider, MediaProviderLanguages};
+use utils::get_base_http_client;
 use utils::{convert_date_to_year, convert_string_to_date};
-
-use crate::app_utils::get_base_http_client;
 
 static URL: &str = "https://api.vndb.org/kana/";
 const METADATA_FIELDS_SMALL: &str = "title,image.url,released,screenshots.url,developers.name";
