+++ crates/utils/application/Cargo.toml
@@ -7,6 +7,10 @@ edition = "2021"
 async-graphql = { workspace = true }
 async-trait = { workspace = true }
 axum = { workspace = true }
+common_utils = { path = "../common" }
 config = { path = "../../config" }
-jwt_service = { path = "../../services/jwt" }
+const-str = { workspace = true }
+dotenvy_macro = { workspace = true }
 file_storage_service = { path = "../../services/file_storage" }
+jwt_service = { path = "../../services/jwt" }
+reqwest = { workspace = true }
