+++ crates/utils/application/src/lib.rs
@@ -7,7 +7,29 @@ use axum::{
     http::{header::AUTHORIZATION, request::Parts, StatusCode},
     Extension, RequestPartsExt,
 };
+use common_utils::PROJECT_NAME;
 use file_storage_service::FileStorageService;
+use reqwest::{
+    header::{HeaderMap, HeaderName, HeaderValue, USER_AGENT},
+    ClientBuilder,
+};
+
+pub const AUTHOR: &str = "ignisda";
+pub const AUTHOR_EMAIL: &str = "ignisda2001@gmail.com";
+#[cfg(debug_assertions)]
+pub const VERSION: &str = dotenvy_macro::dotenv!("APP_VERSION");
+#[cfg(not(debug_assertions))]
+pub const VERSION: &str = env!("APP_VERSION");
+pub const USER_AGENT_STR: &str = const_str::concat!(
+    AUTHOR,
+    "/",
+    PROJECT_NAME,
+    "-v",
+    VERSION,
+    " (",
+    AUTHOR_EMAIL,
+    ")"
+);
 
 pub fn user_id_from_token(token: &str, jwt_secret: &str) -> Result<String> {
     jwt_service::verify(token, jwt_secret)
@@ -58,4 +80,20 @@ pub trait GraphqlRepresentation {
     ) -> Result<Self>
     where
         Self: Sized;
+}
+
+pub fn get_base_http_client(
+    url: &str,
+    headers: Option<Vec<(HeaderName, HeaderValue)>>,
+) -> reqwest::Client {
+    let mut req_headers = HeaderMap::new();
+    req_headers.insert(USER_AGENT, HeaderValue::from_static(USER_AGENT_STR));
+    for (header, value) in headers.unwrap_or_default().into_iter() {
+        req_headers.insert(header, value);
+    }
+    ClientBuilder::new()
+        .default_headers(req_headers)
+        .base_url(url.to_owned())
+        .build()
+        .unwrap()
 }
